+++ components/sceneutil/lightutil.cpp
@@ -73,7 +73,7 @@ namespace SceneUtil
             linearAttenuation = linearMethod == 0 ? linearValue : 0.01f;
             float r = radius * linearRadiusMult;
             if (r > 0.f && (linearMethod == 1 || linearMethod == 2))
-                linearAttenuation = linearValue / std::pow(r, linearMethod);
+                linearAttenuation = linearValue / std::pow(r, static_cast<float>(linearMethod));
         }
 
         if (useQuadratic && (!outQuadInLin || isExterior))
@@ -81,7 +81,7 @@ namespace SceneUtil
             quadraticAttenuation = quadraticMethod == 0 ? quadraticValue : 0.01f;
             float r = radius * quadraticRadiusMult;
             if (r > 0.f && (quadraticMethod == 1 || quadraticMethod == 2))
-                quadraticAttenuation = quadraticValue / std::pow(r, quadraticMethod);
+                quadraticAttenuation = quadraticValue / std::pow(r, static_cast<float>(quadraticMethod));
         }
 
         // If the values are still nonsense, try to at least prevent UB and disable attenuation
